+++ src/components/index.js
@@ -2,4 +2,5 @@ export { default as Button } from './Button';
 export { default as Card } from './Card';
 export { default as Container } from './Container';
 export { default as Footer } from './Footer';
+export { default as H1 } from './H1';
 export { default as NavigationBar } from './Nav';
