+++ src/pages/README.md
@@ -1,0 +1,4 @@
+# Conventions to follow in src/pages
+
+1. Name all page files/ folders in lowercase
+1. Use a separate folder if the page is split into fragment components (sections) and use index file to export the page
