+++ src/pages/home/Members.js
@@ -1,8 +1,7 @@
 import styled from 'styled-components';
-import { Container } from './components';
+import { Card, Container } from '../../components';
 
-import Card from './components/Card';
-import Bheem from './images/man.png';
+import Bheem from '../../images/man.png';
 
 export default function Members() {
 	return (
