+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as Home } from './home';
+export { default as About } from './about';
+export { default as Register } from './register';
